# contracts/openapi/pms-api.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PMS Backend API",
    "version": "1.0.0",
    "description": "Contrat initial publié par pms-backend pour le BFF pms-frontend."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "keycloakBearer": []
    }
  ],
  "paths": {
...
    "/risk/dashboard/{poolId}": {
      "get": {
        "operationId": "getPublishedRiskDashboard",
        "parameters": [
          { "$ref": "#/components/parameters/CorrelationId" },
          { "name": "poolId", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{1,63}$" } }
        ],
        "responses": {
          "200": { "description": "Tableau de risque du dernier run publie", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PublishedRiskDashboard" } } } },
          "default": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/risk/dcr-calculations": {
      "post": {
        "operationId": "calculateDcr",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DcrCalculationCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "DCR calculé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DcrCalculationResult"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/risk/stress-scenarios": {
      "post": {
        "operationId": "runStressScenario",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StressScenarioCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stress scenario exécuté",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StressScenarioResult"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "keycloakBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "CorrelationId": {
        "name": "X-Correlation-ID",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 16,
          "maxLength": 128
        }
      }
    },
    "responses": {
      "Problem": {
        "description": "Erreur normalisée RFC 9457",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "CompositionBucket": {
        "type": "object",
        "required": [
          "bucket",
          "amount"
        ],
        "properties": {
          "bucket": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "PublishedRiskDashboard": {
        "type": "object",
        "required": ["runId", "poolId", "businessDate", "currency", "poolProfitRate", "distributedProfitRate", "yieldGap", "bankMargin", "totalResources", "investedAmount", "averageDurationDays", "assetConcentrationRate", "maturityGaps", "currencyGaps", "perCoverageRate", "irrCoverageRate", "dcrValue", "dcrState", "source", "dataQualityWarnings"],
        "properties": {
          "runId": { "type": "string" },
          "poolId": { "type": "string" },
          "businessDate": { "type": "string", "format": "date" },
          "currency": { "type": "string", "pattern": "^[A-Z]{3}$" },
          "poolProfitRate": { "type": "string" },
          "distributedProfitRate": { "type": "string" },
          "yieldGap": { "type": "string" },
          "bankMargin": { "type": "string" },
          "totalResources": { "type": "string" },
          "investedAmount": { "type": "string" },
          "averageDurationDays": { "type": ["string", "null"] },
          "assetConcentrationRate": { "type": "string" },
          "maturityGaps": { "type": "array", "items": { "$ref": "#/components/schemas/CompositionBucket" } },
          "currencyGaps": { "type": "array", "items": { "$ref": "#/components/schemas/CompositionBucket" } },
          "perCoverageRate": { "type": "string" },
          "irrCoverageRate": { "type": "string" },
          "dcrValue": { "type": ["string", "null"] },
          "dcrState": { "type": ["string", "null"], "enum": ["WITHIN_LIMIT", "BREACH", null] },
          "source": { "type": "string", "const": "LATEST_PUBLISHED_RUN" },
          "dataQualityWarnings": { "type": "array", "items": { "type": "string" } }
        },
        "additionalProperties": false
      },
      "DcrCalculationCommand": {
        "type": "object",
        "required": [
          "poolId",
          "businessDate",
          "currency",
          "capitalDurationAmount",
          "riskWeightedDurationAmount",
          "threshold",
          "formulaVersion",
          "inputChecksumSha256"
        ],
        "properties": {
          "poolId": {
            "type": "string",
            "minLength": 1
          },
          "businessDate": {
            "type": "string",
            "format": "date"
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$"
          },
          "capitalDurationAmount": {
            "type": "string"
          },
          "riskWeightedDurationAmount": {
            "type": "string"
          },
          "threshold": {
            "type": "string"
          },
          "formulaVersion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "inputChecksumSha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          }
        },
        "additionalProperties": false
      },
      "DcrCalculationResult": {
        "type": "object",
        "required": [
          "dcrCalculationId",
          "value",
          "threshold",
          "state"
        ],
        "properties": {
          "dcrCalculationId": {
            "type": "string",
            "format": "uuid"
          },
          "value": {
            "type": "string"
          },
          "threshold": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "WITHIN_LIMIT",
              "BREACH"
            ]
          }
        },
        "additionalProperties": false
      },
      "StressShock": {
        "type": "object",
        "required": [
          "bucket",
          "basisPoints"
        ],
        "properties": {
          "bucket": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_-]{1,63}$"
          },
          "basisPoints": {
            "type": "integer",
            "minimum": -10000,
            "maximum": 10000
          }
        },
        "additionalProperties": false
      },
      "StressScenarioCommand": {
        "type": "object",
        "required": [
          "scenarioCode",
          "businessDate",
          "currency",
          "baseAmount",
          "shocks",
          "engineVersion",
          "inputChecksumSha256"
        ],
        "properties": {
          "scenarioCode": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_-]{1,63}$"
          },
          "businessDate": {
            "type": "string",
            "format": "date"
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$"
          },
          "baseAmount": {
            "type": "string"
          },
          "shocks": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/StressShock"
            }
          },
          "engineVersion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "inputChecksumSha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          }
        },
        "additionalProperties": false
      },
      "StressResultLine": {
        "type": "object",
        "required": [
          "bucket",
          "basisPoints",
          "stressedAmount",
          "impactAmount"
        ],
        "properties": {
          "bucket": {
            "type": "string"
          },
          "basisPoints": {
            "type": "integer"
          },
          "stressedAmount": {
            "type": "string"
          },
          "impactAmount": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "StressScenarioResult": {
        "type": "object",
        "required": [
          "stressScenarioId",
          "state",
          "results"
        ],
        "properties": {
          "stressScenarioId": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "const": "COMPLETED"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StressResultLine"
            }
          }
        },
        "additionalProperties": false
      },
...
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "correlationId": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    }
  }
}